Add instruction class and hook up to CFU
Common functionality is moved into util.py.

cfu.py is now just an example of implementing a CFU with a few
instructions and tests for those instructions.

CombTest and SyncTest are replaced with a single abstract test class
TestBase which works with both sync and non sync modules. This allows
instruction tests to subclass InstructionTestBase, without needing
separate class hierarchies for comb/sync instructions.

This commit was codeveloped by avg and dml. Sometimes working together
and sometimes not.