Add nmigen as a third party submodule
Also put in place a mechanism for automatically locating such
dependencies. So you don't need a virtual environment or to "install"
each dependency. Just running "make soc" in a standard shell now works.

For now I've left litex as it was - in soc/deps/litex. It's certainly
possible to do the same thing with litex. i.e. add it as a submodule in
third_party/python together with any litex dependencies that are needed.
This is what I did in the HPS repository.