Move all third_party python dependencies to third_party/python
Several reasons for doing this:
- I think google requires that all third_party code be in third_party
- `source environment` will now give you a python path that includes
litex and other dependencies. This makes it possible to easily navigate
sources in vscode, since it now knows where all the dependencies are
located, so you can ctrl+click on things to jump to definitions.
- It allows lxbuildenv to be removed
- It makes things more consistent - building the cfu.v and building the
soc now both use the same way of locating python packages.