Set io_bus_rsp_payload_response_ok=1
I initially tried setting reset=1 on the Signal. This made the nmigen
test pass, however it wasn't sufficient to make the CFU actually work.
Looking at gtkwave, it seemed that the signal still ended up low. I'm
not sure why the discrepancy between the two, or if there's something we
can do about it (e.g. can we make the test more accurate and detect that
reset isn't sufficient)