Move util.py into a separate python library
I called this nmigen_cfu for now, since it's primarily a library for
writing CFUs with nMigen. It also includes some of the testing utils
stuff from Alan's original util.py. That could be separated out to a
different library at some point if we wanted. For now, I've just left it
all together.