Replace filter_x and filter_y with 0:
In ConvPerChannelSpecialized, replace filter_x and filter_y with 0. This
can be done since we only enter ConvPerChannelSpecialized when
filter_height and filter_width are 1 and from the for loop, filter_y <
filter_height. Therefore, filter_y can be replaced with 0 and this is
the same for filter_x.

This produces a total acceleration of 51.86% and an incremental
acceleration of 0.04%.

Signed-off-by: Rachel Sugrono <rsugrono@google.com>