New Make strategy
Slightly revised build method. This method:

- uses common instead of tflm_harness
- compiles tflm as part of the binary rather than as a separate library
- allows any part of the library or the harness to be overlaid.

The implentation is mostly in proj/proj2.mk, which will eventually
replace proj/proj.mk.  The Makefile in the repo root is not used in this
new strategy and will eventually be deleted.

I have tested this with renode, but not yet on an Arty, since I don't
have a working one handy.  Only proj_template has been changed over at
this point.

Signed-off-by: Alan Green <avg@google.com>