hps: make period constraint account for imprecise internal oscillator
On the CrossLink-NX, the internal oscillator nominally runs at 450MHz but may
actually run at 405-495MHz. On my board it appears to run at ~488MHz.

If the oscillator produces a clock signal faster than we are expecting, it
means we actually have less time for signals to propagate. We need to specify a
period constraint accounting for the worst case (10% faster than nominal)
otherwise there is a chance that the design will meet timing, but when run on a
real device signals will not propagate fast enough resulting in strange
misbehaviour.