hps: add an option to populate integrated ROM
This will replace the Litex BIOS which is normally loaded as the
integrated ROM contents. The option is named --integrated-rom-init for
consistency with SoCCore which defines the same option.

This new option, alongside the support for executing out of LRAM in the
previous commit, makes it possible to tweak the Makefiles so that the
CFU Playground application runs entirely from memory -- although the
Makefiles are left as is in this commit.