report "ticks" to the tflite-micro profiler as cycles / 1024
In tflite-micro the profiling code uses int32_t to represent ticks,
presumably because it was intended to be fed from a relatively low
precision timer source and not the raw system clock cycle counter.

The HPS model (unaccelerated) has one layer which takes more than 2**31
cycles and so it was being displayed in the profile as taking negative
time.

Report "ticks" as clock cycles / 1024, instead of raw cycles. The choice
of divisor is arbitrary but 1024 can be implemented in a single shift
instruction and still makes the numbers easy to interpret (humans just
have to think of it as "kilocycles").

Also, for completeness sake, use the full 64-bit cycle counter in the
unlikely event that a single profiling event takes more than 2**32
cycles. This implementation of get_mcycle() is taken from OpenTitan:
https://github.com/lowRISC/opentitan/blob/d9dc048680d89bd630ffae673103509e28017709/sw/device/lib/runtime/ibex.h#L28
based on the recommendation in the RISC-V Instruction Set Manual.