Make CI more flexible for new projects
With this change CI jobs are generated based on additional files:
1. `.github/workflows/projects_to_test.txt` - list of project
names to test in CI
2. `.github/workflows/supported_targets.txt` - list of target
names that are supported by both LiteX Boards and CFU-Playground
3. `proj/<proj_name>/ci/ci_exclude_targets.txt` - list of target
names that shouldn't be tested in CI for a specific project
4. `proj/<proj_name>/ci/ci_build_params.txt` - additional parame-
ters that should be passed to build a specific project
5. `proj/<proj_name>/renode/<renode_file>` - Renode files
(`<target>.resc`, `<target>.repl`, `<target>.robot`) that should
not be auto-generated. If file with target names exist, it will be
used without applying any changes.

Files outside `.github` directory are optional and not providing
them will yield using default values.

Signed-off-by: Robert Szczepanski <rszczepanski@antmicro.com>