hps: add platform definition for proto2 board
The new hps_proto2_platform.py is a copy of the existing
hps_proto0_platform.py, with a few pin assignments adjusted to match the
different layout on the proto2 board.

Use the new platform by default for HPS.

The original hps_proto0_platform.py is left unused in this commit but if
necessary we can switch back to using proto0 by adjusting the import in
hps_soc.py.

Signed-off-by: Dan Callaghan <dcallagh@google.com>