hps_accel: implement filter store registers
This is largely a copy-paste from the input store implementation, with
constants and memory depth tweaked, as a starting point. We expect the
two implementations will start to diverge after this (particularly since
the filter store will need to be shrunk to fit inside the fewer EBRAMs
on HPS).

The tests in test_filter_store.py are independently implemented, which
is why they look quite different to test_input_store.py.

Signed-off-by: Dan Callaghan <dcallagh@google.com>