hps_accel: fix off-by-one error with max_index signal width
This signal was called max_index (implying its value should range up to
N-1 for N input words) but it was actually being used as num_words (its
value would range up to N for N input words).

The logic using the signal was correct, but the signal's width was
wrongly calculated. If num_words was set equal to the maximum capacity
of the input store, the max_index signal would wrap around to 0 instead
of N.

Fix the logic to make the max_index signal's meaning match its name
(ranging up to N-1).

Signed-off-by: Dan Callaghan <dcallagh@google.com>