hps_accel: disconnect store outputs from their registers
We keep the registers, and pipe through the payload, but we ignore the
flow control signals. Effectively this means the registers are just
"spying" on the data in the stream but not participating in flow
control. This is in preparation for letting the macc unit control the
flow instead.

As a consequence, we have to remove the invalidation logic from these
registers, meaning that they can return out-of-date values. We need to
insert some nops to work around the timing issues. These registers are
only needed for debugging and testing so this seems harmless enough.

Signed-off-by: Dan Callaghan <dcallagh@google.com>