hps_accel: stores: don't emit first values until 'next' strobe
This changes the meaning of the 'next' strobe for the input store and
filter store.

Previously, the stores emitted their first value as soon as the write
sequence finished. Then they emitted the next value for every 'next'
strobe. That meant if we wanted them to emit N values there must be N-1
strobes on 'next'.

Instead, the store will only emit a new value while 'next' is asserted.
That means the software can send N 'next' signals to trigger N values.
It will help to simplify the logic when we need to start coordinating
long sequences of values from the stores, potentially being reset at
different times.

Signed-off-by: Dan Callaghan <dcallagh@google.com>