hps_accel: ensure every macc output value is read exactly once
Change the behaviour of REG_MACC_OUT so that the CPU will block until
the register contains a valid value, *and* the register's input stream
will block until the CPU has read out the existing value and invalidated
the register contents.

This lets us remove all nops that were accounting for pipeline delays.
This also prepares for a future commit where we will make the CPU wait
while the macc unit runs through a longer sequence of operations.

Also update the software CFU to mimic the same behaviour. This helped me
to catch a problem in how blocks_test.cc was polluting the macc with
leftover values.

Signed-off-by: Dan Callaghan <dcallagh@google.com>