hps_accel/conv.h: find output parameter ranges
Allows the finding of ranges of data used in post processing.

The results of running with this code on, for the hps model were:

multiplier min: 1074667989 (0x400e21d5) max: 2143413248 (0x7fc1e400)
     shift min:        -10 (0xfffffff6) max:         -6 (0xfffffffa)
      bias min:     -18945 (0xffffb5ff) max:      19445 (0x00004bf5)

This commit turn SHOW_CONV_PARAMS off so these values will not be
printed in future. It also puts the output behind a "#if 0" so a code
change will be needed to turn it back when required.

Signed-off-by: Alan Green <avg@google.com>