hps_accel: remove store "spying" registers and associated tests
Since commit 87f3de00, the REG_INPUT_{1..4} and REG_FILTER_{1..4}
registers are just "spying" on the streams coming out of the stores, for
testing purposes. They don't participate in flow control.

But as we improve the stores to stream out values with less latency,
it's no longer possible for the CPU to observe each individual set of
values going through these registers. The stores will be capable of
updating the registers faster than the CPU can read out all four
registers.

Remove the registers and all the associated test cases, since they are
now more of a hindrance than a help.

Signed-off-by: Dan Callaghan <dcallagh@google.com>