hps_accel: update FilterStore implementation to match InputStore
FilterStore originally started out as a copy-paste of InputStore. We
expected them to diverge in behaviour and implementation, but they
haven't yet (not intentionally).

Replace the old FilterStore implementation a fresh copy-paste of the
latest InputStore using WideReadMemory.

This fixes some outstanding issues with reset behaviour for FilterStore.
We lose the ability to stream values with zero-cycle latency but that
will be addressed in a follow-up commit.

Signed-off-by: Dan Callaghan <dcallagh@google.com>