hps_accel: introduce FlowRestrictor
This component sits in between an upstream and downstream and controls
the flow of packets through the stream. It keeps the stream paused until
it is instructed to release a specific number of packets in a batch,
through its 'release' endpoint.

We will place this in between our stores and the operand buffer, in
order to let the CPU send N sets of values through the macc pipeline.

Signed-off-by: Dan Callaghan <dcallagh@google.com>