hps_accel: run all macc operations for each output value in a batch
Rather than looping in software for N iterations, each time telling the
stores to release one set of values and then collecting the macc result,
we will tell the stores to release N sets of values and then wait for
N results to be ready.

This is a preparatory step towards accumulating all results in gateware
so that no software loop is required at all.

To accomplish this, we remove the 'next' strobe from the stores. They
now obey back pressure on their output stream but otherwise will read
out values continuously. The current implementation takes 3 cycles for
every output, which is not optimal but I got bogged down trying to
improve this by making the logic even more complicated so this should be
considered just a starting point.

In between each store and the operand buffer, we interpose
a FlowRestrictor. The REG_FILTER_INPUT_NEXT register now controls these
two FlowRestrictors. Each time a value is written to
REG_FILTER_INPUT_NEXT it causes both FlowRestrictors to release the
given number of values onto their streams.

The software CFU implementation is updated to match the new behaviour of
REG_FILTER_INPUT_NEXT. It doesn't simulate the pipelines flowing inside
the gateware, but rather it just asserts the expected usage pattern:
after releasing N sets of values, the software should collect N macc
results.

Signed-off-by: Dan Callaghan <dcallagh@google.com>