hps_accel: single Conv2D test
Tests a single Conv2D op, based on actual data from the HPS model. A
single layer is easier to work with in some kinds of debugging.

Includes code to dump a layer's input and output data to the console, as
well as infrastructure to use that dumped data in a test.

This test will not work on the proto2 board due to memory size
constraints. In future, we may wish to exclude it from the proto2
binary, but there is currently no simple mechanism to do so.

Signed-off-by: Alan Green <avg@google.com>