hps_accel: include proj.mk before checking $(TARGET)
If we don't supply an explicit TARGET to make, it should be set to the default
value in proj.mk, which is digilent_arty. However, this assignment of the
default value is currently done _after_ the ifeq in hps_accel/Makefile, so the
clock frequency isn't reduced to 75MHz even though it should.

Signed-off-by: Jakub Piecuch <jpiecuch@antmicro.com>