gen2/sample_conv_1: Add test data and framework
A pickle file of data from a real evaluation of the 09_20 hps model.
Includes the script to produce the pickle file and functions to read and
write the pickle file.

We use a pickle file to hold data as the source code version of the raw
data is inconveniennt to use in an IDE.

Signed-off-by: Alan Green <avg@google.com>