hps_accel/conv_accel_gen_2.cc: fix gating
More strictly controls which Conv2D ops are accelerated. In particular
this change only allows Conv2D ops where input and output depths are
exactly 16. With this change, golden tests now pass for the 09_20 model.

In future, the accelerator will be generalised further to allow more
Conv2D ops to be accelerated.

Signed-off-by: Alan Green <avg@google.com>