Update Setup Guide for Arty A7 using Symbiflow
Add note about symbiflow route and place tools not completing if
unable to meet timing. This will be likely to occur using the
suggested Arty A7 when completing the Test Run in Step 6 of the
Setup Guide. As this can be confusing for new users, this change
suggests the user can add the command to lower the target clock
rate when completing this step if required.

Signed-off-by: Algy Tynan <algy@tynan.io>