Rebuild local VexRiscv verilog @aed4e355.
Now at tip of master of the VexRiscv.
The issue of #596 is avoided by configuring the CfuPlugin
with "withEnable = false".

This also incorporates a change to add the attribute
(* no_rw_check *) on the block memories in VexRiscv --
regfile and caches.   This disables circuitry for a check
added by Yosys that we don't need, according to Charles P.
Having the check consumes resources, particularly noticeable
when targetting ice40.

Signed-off-by: Tim Callahan <tcal@google.com>